Removed struct definition from the .h file. Converted macros to functions
